agrega excepciones para mejor manejo de errores

+++ src/main/java/pt/fmbp/soiapbackend/controller/HoraAtencionController.java
@@ -6,10 +6,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.*;
 import pt.fmbp.soiapbackend.entity.HoraAtencion;
-import pt.fmbp.soiapbackend.exception.APICustomException;
 import pt.fmbp.soiapbackend.exception.AppointmentConflictException;
+import pt.fmbp.soiapbackend.exception.InvalidIdException;
+import pt.fmbp.soiapbackend.exception.InvalidParameterException;
+import pt.fmbp.soiapbackend.exception.ResourceNotFoundException;
 import pt.fmbp.soiapbackend.service.IHoraAtencionService;
 
+import javax.validation.Valid;
 import java.text.ParseException;
 import java.time.LocalDate;
 import java.util.List;
@@ -24,7 +27,7 @@ public class HoraAtencionController {
     // Crear un nuevo horario de atención
     @Secured({"ROLE_PSICOLOGO_TRATANTE", "ROLE_COLABORADOR"})
     @PostMapping("")
-    public ResponseEntity<HoraAtencion> createHoraAtencion (@RequestBody HoraAtencion horaAtencion) throws ParseException {
+    public ResponseEntity<HoraAtencion> createHoraAtencion (@Valid @RequestBody HoraAtencion horaAtencion) throws ParseException {
         if (horaAtencion != null) {
             HoraAtencion horaAtencionCreada = horaAtencionService.saveHoraAtencion(horaAtencion);
 
@@ -35,25 +38,31 @@ public class HoraAtencionController {
             else
                 return new ResponseEntity<>(horaAtencionCreada, HttpStatus.CREATED);
         } else
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
     // Obtener un horario de atención específico
     @Secured({"ROLE_PSICOLOGO_TRATANTE", "ROLE_COLABORADOR"})
     @GetMapping("/{idAtencion}")
     public ResponseEntity<HoraAtencion> getHoraAtencion (@PathVariable(value = "idAtencion") Long idAtencion) {
+        if (idAtencion == null || idAtencion == 0) {
+            throw new InvalidIdException("El ID del horario de atención ingresado no es válido.");
+        }
+
         HoraAtencion horaAtencionEncontrada = horaAtencionService.getHoraAtencion(idAtencion);
 
         if (horaAtencionEncontrada != null) {
             return new ResponseEntity<>(horaAtencionEncontrada, HttpStatus.OK);
         }
-         else return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+         else throw new ResourceNotFoundException("El horario de atención no ha sido encontrado.");
     }
 
     // Obtener horarios de atención por fecha
     @Secured({"ROLE_PSICOLOGO_TRATANTE", "ROLE_COLABORADOR"})
     @GetMapping("/get/{fecha}")
     public ResponseEntity<List<HoraAtencion>> getHoraAtencionPorFecha(@PathVariable(value = "fecha") String fechaAtencion) {
+        if (fechaAtencion == null) throw new InvalidParameterException("Debe ingresar una fecha para realizar la búsqueda.");
+
         LocalDate localDate = LocalDate.parse(fechaAtencion);
         List<HoraAtencion> horarios = horaAtencionService.getHorasPorFecha(localDate);
 
@@ -63,7 +72,7 @@ public class HoraAtencionController {
     // Actualizar una hora de atención
     @Secured({"ROLE_PSICOLOGO_TRATANTE", "ROLE_COLABORADOR"})
     @PutMapping("/update/{idAtencion}")
-    public ResponseEntity<HoraAtencion> updateHoraAtencion (@RequestBody HoraAtencion horaAtencion,
+    public ResponseEntity<HoraAtencion> updateHoraAtencion (@Valid @RequestBody HoraAtencion horaAtencion,
                                                             @PathVariable(value = "idAtencion") Long idAtencion) {
         HoraAtencion horaAtencionActualizada = horaAtencionService.updateHoraAtencion(horaAtencion, idAtencion);
 
@@ -77,6 +86,8 @@ public class HoraAtencionController {
     @DeleteMapping("/delete/{idAtencion}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteHoraAtencion (@PathVariable(value = "idAtencion") Long idAtencion) {
+        if (idAtencion == null || idAtencion == 0) throw new InvalidIdException("El ID de la atención proporcionado no es válido.");
+
         horaAtencionService.deleteHoraAtencion(idAtencion);
     }
 }
